. function createRepl
