+++ src/f/createRepl.js
@@ -1,0 +1,21 @@
+const Parinfer = require('parinfer'); // Assuming you have the parinfer module installed
+
+function createRepl(rl) {
+  rl.on('line', (line) => {
+    // Use Parinfer to process the line
+    const processedLine = Parinfer.indentMode(line, { // Use indentMode or parenMode as needed
+      // Parinfer options
+    });
+
+    // Display the processed line (or handle errors/changes)
+    console.log(processedLine);
+  });
+
+  rl.on('close', () => {
+    console.log('Readline interface closed.');
+  });
+
+  return this;
+}
+
+module.exports = createRepl;
